<FCV2> Increased PWM range + tuned initial PID value
Changes:
1. Increased PWM range from 0~100 to 0~1000
2. Tuned intial PID value to at least make sense
3. Fixed motor control direction.

Test & Result:
Tested on drone. Seems fine without propeller.

+++ FlightController/FlightControllerV2/Inc/pwm.h
@@ -19,8 +19,7 @@ void PWM_Start();
 void PWM_Stop();
 // bool PWM_SetFrequency();
 
-// dutyCycle : between 0 ~ 100
-void PWM_SetDutyCycle(PWMChannelType channel, uint8_t dutyCycle);
+void PWM_SetDutyCycle(PWMChannelType channel, int dutyCycle);
 
 #ifdef __cplusplus
 }
